Renaming video question to specialized-video.question to improve discoverability Extracting common file type logic to a separate file DRY principle for pdf file viewer and file viewer
