+++ apps/endatix-hub/features/submissions/ui/details/submission-details.tsx
@@ -8,7 +8,7 @@ import AnswerViewer from '../answers/answer-viewer';
 import { SectionTitle } from '@/components/headings/section-title';
 import { BackToSubmissionsButton } from './back-to-submissions-button';
 import { registerSpecializedQuestion } from '@/lib/questions';
-import { SpecializedVideo } from '@/lib/questions/video/specialized-video';
+import { SpecializedVideo } from '@/lib/questions/video/specialized-video.question';
 
 registerSpecializedQuestion(SpecializedVideo);
 async function SubmissionDetails({
